Create rock paper scissors game
This consists of a bare-bones html document with a seperate linked JS file.
The game is meant to be played entirely in the console and is invoked with game().
This plays a best-three-out-of-five series where the user inputs their choice and the
computer randomly selects a choice. If either the player or the computer get 3 wins
before five rounds is played, the game is over. Ties mean a replay.